feat(auth): add customer validation and enhance authentication flow

+++ backend/bank-auth-service/src/main/java/com/bank/auth/dto/RegisterResponse.java
@@ -1,30 +1,39 @@
 package com.bank.auth.dto;
 
 public class RegisterResponse {
-	private Long customerId;
-	private String message;
-	
-	public RegisterResponse() {
-		
-	}
-	public RegisterResponse(Long customerId, String message) {
-		this.customerId = customerId;
-		this.message = message;
-		
-	}
-	public Long getCustomerId() {
-		return customerId;
-	}
-	public void setCustomerId(Long customerId) {
-		this.customerId = customerId;
-	}
-	public String getMessage() {
-		return message;
-	}
-	public void setMessage(String message) {
-		this.message = message;
-	}
-	
-	
+
+    private Integer customerId;
+    private String message;
+
+
+    public RegisterResponse() {
+
+    }
+
+
+    public RegisterResponse(Integer customerId, String message) {
+        this.customerId = customerId;
+        this.message = message;
+    }
+
+
+    public Integer getCustomerId() {
+        return customerId;
+    }
+
+
+    public void setCustomerId(Integer customerId) {
+        this.customerId = customerId;
+    }
+
+
+    public String getMessage() {
+        return message;
+    }
+
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
 
 }
